+++ project/GeneratorAxis.scala
@@ -1,0 +1,3 @@
+import sbt._
+
+case class GeneratorAxis(idSuffix: String, directorySuffix: String) extends VirtualAxis.WeakAxis