+++ scalapb-runtime/src/main/scala/scalapb/options/ScalaPbOptions.scala
@@ -79,6 +79,8 @@ package scalapb.options
   *   opt out from transformations inherited through package-scoped options.
   * @param getters
   *   If true, getters will be generated.
+  * @param scala3Sources
+  *   Generate sources that are compatible with Scala 3
   * @param testOnlyNoJavaConversions
   *   For use in tests only. Inhibit Java conversions even when when generator parameters
   *   request for it.
@@ -112,6 +114,7 @@ final case class ScalaPbOptions(
     fieldTransformations: _root_.scala.Seq[scalapb.options.FieldTransformation] = _root_.scala.Seq.empty,
     ignoreAllTransformations: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None,
     getters: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None,
+    scala3Sources: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None,
     testOnlyNoJavaConversions: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None,
     unknownFields: _root_.scalapb.UnknownFieldSet = _root_.scalapb.UnknownFieldSet.empty
     ) extends scalapb.GeneratedMessage with scalapb.lenses.Updatable[ScalaPbOptions] with _root_.scalapb.ExtendableMessage[ScalaPbOptions] {
@@ -226,6 +229,10 @@ final case class ScalaPbOptions(
       if (getters.isDefined) {
         val __value = getters.get
         __size += _root_.com.google.protobuf.CodedOutputStream.computeBoolSize(27, __value)
+      };
+      if (scala3Sources.isDefined) {
+        val __value = scala3Sources.get
+        __size += _root_.com.google.protobuf.CodedOutputStream.computeBoolSize(28, __value)
       };
       if (testOnlyNoJavaConversions.isDefined) {
         val __value = testOnlyNoJavaConversions.get
@@ -362,6 +369,10 @@ final case class ScalaPbOptions(
         val __m = __v
         _output__.writeBool(27, __m)
       };
+      scala3Sources.foreach { __v =>
+        val __m = __v
+        _output__.writeBool(28, __m)
+      };
       testOnlyNoJavaConversions.foreach { __v =>
         val __m = __v
         _output__.writeBool(999, __m)
@@ -457,6 +468,9 @@ final case class ScalaPbOptions(
     def getGetters: _root_.scala.Boolean = getters.getOrElse(true)
     def clearGetters: ScalaPbOptions = copy(getters = _root_.scala.None)
     def withGetters(__v: _root_.scala.Boolean): ScalaPbOptions = copy(getters = Option(__v))
+    def getScala3Sources: _root_.scala.Boolean = scala3Sources.getOrElse(false)
+    def clearScala3Sources: ScalaPbOptions = copy(scala3Sources = _root_.scala.None)
+    def withScala3Sources(__v: _root_.scala.Boolean): ScalaPbOptions = copy(scala3Sources = Option(__v))
     def getTestOnlyNoJavaConversions: _root_.scala.Boolean = testOnlyNoJavaConversions.getOrElse(false)
     def clearTestOnlyNoJavaConversions: ScalaPbOptions = copy(testOnlyNoJavaConversions = _root_.scala.None)
     def withTestOnlyNoJavaConversions(__v: _root_.scala.Boolean): ScalaPbOptions = copy(testOnlyNoJavaConversions = Option(__v))
@@ -491,6 +505,7 @@ final case class ScalaPbOptions(
         case 25 => fieldTransformations
         case 26 => ignoreAllTransformations.orNull
         case 27 => getters.orNull
+        case 28 => scala3Sources.orNull
         case 999 => testOnlyNoJavaConversions.orNull
       }
     }
@@ -524,6 +539,7 @@ final case class ScalaPbOptions(
         case 25 => _root_.scalapb.descriptors.PRepeated(fieldTransformations.iterator.map(_.toPMessage).toVector)
         case 26 => ignoreAllTransformations.map(_root_.scalapb.descriptors.PBoolean(_)).getOrElse(_root_.scalapb.descriptors.PEmpty)
         case 27 => getters.map(_root_.scalapb.descriptors.PBoolean(_)).getOrElse(_root_.scalapb.descriptors.PEmpty)
+        case 28 => scala3Sources.map(_root_.scalapb.descriptors.PBoolean(_)).getOrElse(_root_.scalapb.descriptors.PEmpty)
         case 999 => testOnlyNoJavaConversions.map(_root_.scalapb.descriptors.PBoolean(_)).getOrElse(_root_.scalapb.descriptors.PEmpty)
       }
     }
@@ -562,6 +578,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     val __fieldTransformations: _root_.scala.collection.immutable.VectorBuilder[scalapb.options.FieldTransformation] = new _root_.scala.collection.immutable.VectorBuilder[scalapb.options.FieldTransformation]
     var __ignoreAllTransformations: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None
     var __getters: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None
+    var __scala3Sources: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None
     var __testOnlyNoJavaConversions: _root_.scala.Option[_root_.scala.Boolean] = _root_.scala.None
     var `_unknownFields__`: _root_.scalapb.UnknownFieldSet.Builder = null
     var _done__ = false
@@ -623,6 +640,8 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
           __ignoreAllTransformations = Option(_input__.readBool())
         case 216 =>
           __getters = Option(_input__.readBool())
+        case 224 =>
+          __scala3Sources = Option(_input__.readBool())
         case 7992 =>
           __testOnlyNoJavaConversions = Option(_input__.readBool())
         case tag =>
@@ -660,6 +679,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
         fieldTransformations = __fieldTransformations.result(),
         ignoreAllTransformations = __ignoreAllTransformations,
         getters = __getters,
+        scala3Sources = __scala3Sources,
         testOnlyNoJavaConversions = __testOnlyNoJavaConversions,
         unknownFields = if (_unknownFields__ == null) _root_.scalapb.UnknownFieldSet.empty else _unknownFields__.result()
     )
@@ -695,6 +715,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
         fieldTransformations = __fieldsMap.get(scalaDescriptor.findFieldByNumber(25).get).map(_.as[_root_.scala.Seq[scalapb.options.FieldTransformation]]).getOrElse(_root_.scala.Seq.empty),
         ignoreAllTransformations = __fieldsMap.get(scalaDescriptor.findFieldByNumber(26).get).flatMap(_.as[_root_.scala.Option[_root_.scala.Boolean]]),
         getters = __fieldsMap.get(scalaDescriptor.findFieldByNumber(27).get).flatMap(_.as[_root_.scala.Option[_root_.scala.Boolean]]),
+        scala3Sources = __fieldsMap.get(scalaDescriptor.findFieldByNumber(28).get).flatMap(_.as[_root_.scala.Option[_root_.scala.Boolean]]),
         testOnlyNoJavaConversions = __fieldsMap.get(scalaDescriptor.findFieldByNumber(999).get).flatMap(_.as[_root_.scala.Option[_root_.scala.Boolean]])
       )
     case _ => throw new RuntimeException("Expected PMessage")
@@ -753,6 +774,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     fieldTransformations = _root_.scala.Seq.empty,
     ignoreAllTransformations = _root_.scala.None,
     getters = _root_.scala.None,
+    scala3Sources = _root_.scala.None,
     testOnlyNoJavaConversions = _root_.scala.None
   )
   /** Whether to apply the options only to this file, or for the entire package (and its subpackages)
@@ -790,7 +812,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     
     @SerialVersionUID(0L)
     final case class Unrecognized(unrecognizedValue: _root_.scala.Int) extends OptionsScope(unrecognizedValue) with _root_.scalapb.UnrecognizedEnum
-    lazy val values = scala.collection.immutable.Seq(FILE, PACKAGE)
+    lazy val values: scala.collection.immutable.Seq[ValueType] = scala.collection.immutable.Seq(FILE, PACKAGE)
     def fromValue(__value: _root_.scala.Int): OptionsScope = __value match {
       case 0 => FILE
       case 1 => PACKAGE
@@ -834,7 +856,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     
     @SerialVersionUID(0L)
     final case class Unrecognized(unrecognizedValue: _root_.scala.Int) extends EnumValueNaming(unrecognizedValue) with _root_.scalapb.UnrecognizedEnum
-    lazy val values = scala.collection.immutable.Seq(AS_IN_PROTO, CAMEL_CASE)
+    lazy val values: scala.collection.immutable.Seq[ValueType] = scala.collection.immutable.Seq(AS_IN_PROTO, CAMEL_CASE)
     def fromValue(__value: _root_.scala.Int): EnumValueNaming = __value match {
       case 0 => AS_IN_PROTO
       case 1 => CAMEL_CASE
@@ -1486,6 +1508,8 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     def optionalIgnoreAllTransformations: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Option[_root_.scala.Boolean]] = field(_.ignoreAllTransformations)((c_, f_) => c_.copy(ignoreAllTransformations = f_))
     def getters: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Boolean] = field(_.getGetters)((c_, f_) => c_.copy(getters = Option(f_)))
     def optionalGetters: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Option[_root_.scala.Boolean]] = field(_.getters)((c_, f_) => c_.copy(getters = f_))
+    def scala3Sources: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Boolean] = field(_.getScala3Sources)((c_, f_) => c_.copy(scala3Sources = Option(f_)))
+    def optionalScala3Sources: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Option[_root_.scala.Boolean]] = field(_.scala3Sources)((c_, f_) => c_.copy(scala3Sources = f_))
     def testOnlyNoJavaConversions: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Boolean] = field(_.getTestOnlyNoJavaConversions)((c_, f_) => c_.copy(testOnlyNoJavaConversions = Option(f_)))
     def optionalTestOnlyNoJavaConversions: _root_.scalapb.lenses.Lens[UpperPB, _root_.scala.Option[_root_.scala.Boolean]] = field(_.testOnlyNoJavaConversions)((c_, f_) => c_.copy(testOnlyNoJavaConversions = f_))
   }
@@ -1516,6 +1540,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
   final val FIELD_TRANSFORMATIONS_FIELD_NUMBER = 25
   final val IGNORE_ALL_TRANSFORMATIONS_FIELD_NUMBER = 26
   final val GETTERS_FIELD_NUMBER = 27
+  final val SCALA3_SOURCES_FIELD_NUMBER = 28
   final val TEST_ONLY_NO_JAVA_CONVERSIONS_FIELD_NUMBER = 999
   def of(
     packageName: _root_.scala.Option[_root_.scala.Predef.String],
@@ -1545,6 +1570,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     fieldTransformations: _root_.scala.Seq[scalapb.options.FieldTransformation],
     ignoreAllTransformations: _root_.scala.Option[_root_.scala.Boolean],
     getters: _root_.scala.Option[_root_.scala.Boolean],
+    scala3Sources: _root_.scala.Option[_root_.scala.Boolean],
     testOnlyNoJavaConversions: _root_.scala.Option[_root_.scala.Boolean]
   ): _root_.scalapb.options.ScalaPbOptions = _root_.scalapb.options.ScalaPbOptions(
     packageName,
@@ -1574,6 +1600,7 @@ object ScalaPbOptions extends scalapb.GeneratedMessageCompanion[scalapb.options.
     fieldTransformations,
     ignoreAllTransformations,
     getters,
+    scala3Sources,
     testOnlyNoJavaConversions
   )
   // @@protoc_insertion_point(GeneratedMessageCompanion[scalapb.ScalaPbOptions])