+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.11.14]
 - Repeated extensions can now use arbitrary iterable types or any collection type there is a
   CollectionAdapter for (#1509)
+- Introduced new generator parameter `scala3_sources` (also available via
+  scalapb.proto) to generate sources that are compatible with Scala 3 (#1576)
 
 ## [0.11.13]
 - Added input and output message type for method descriptor (#1503)